requirements.txt
python-telegram-bot==20.5
bcrypt
firebase-admin
python-dotenv
requests
